requirements-dev.txt
-r requirements.txt
pytest==5.4.1
coverage==5.1
flake8==3.8.0a2
pep8-naming==0.10.0
freezegun==0.3.15
pytest-cov==2.8.1
codecov==2.0.22
mccabe==0.6.1
zest.releaser==6.20.1
requests-mock==1.8.0
ipdb==0.13.2
tox==3.15.0
